requirements.txt
asgiref==3.5.2
boto3==1.26.7
botocore==1.29.7
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
celery
coreschema==0.0.4
Django==3.2
django-cors-headers==3.13.0
django-easy-select2==1.5.7
django-filter==2.1.0
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
openapi-codec==1.3.2
Pillow==9.3.0
pycodestyle==2.10.0
PyJWT==1.7.1
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2022.2.1
razorpay==1.3.0
requests==2.28.1
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
tomli==2.0.1
tzdata>=2022.2
uritemplate==4.1.1
urllib3==1.26.12
instagrapi==1.17.10
celery==5.3.1